+++ src-tauri/src/vojo/get_base_config_response.rs
@@ -6,14 +6,14 @@ pub struct GetBaseConnectionResponse {
 }
 #[derive(Deserialize, Serialize)]
 pub struct GetBaseConnectionResponseItem {
-    pub base_config_id: i32,
+    pub base_config_id: u64,
     pub connection_name: String,
     pub connection_type: i32,
     pub description: String,
 }
 #[derive(Deserialize, Serialize)]
 pub struct GetBaseConnectionByIdResponse {
-    pub base_config_id: i32,
+    pub base_config_id: u64,
     pub connection_name: String,
     pub connection_json: String,
     pub connection_type: i32,
